feat: Add GallerySection.tsx
